backend/requirements.txt
flask
flask-cors
gunicorn
joblib
pandas
seaborn
numpy
scikit-learn
sentence-transformers
faiss-cpu
PyMuPDF
requests
python-dotenv
deep-translator
arabic-reshaper
python-bidi
openrouter
gunicorn